fixed naming conventions

+++ aspens/src/commands/trading/balance.rs
@@ -16,7 +16,7 @@ use crate::commands::config::config_pb::{Configuration, GetConfigResponse};
 struct TokenInfo {
     symbol: String,
     name: String,
-    decimals: i32,
+    decimals: u32,
     /// Map of chain_id -> (chain_network, token_address, contract_address)
     chain_locations: HashMap<u32, ChainLocation>,
 }
@@ -169,14 +169,14 @@ async fn query_token_balance_on_chain(
 }
 
 /// Format balance with decimals for human-readable display
-fn format_balance_with_decimals(balance_str: &str, decimals: i32) -> String {
+fn format_balance_with_decimals(balance_str: &str, decimals: u32) -> String {
     if balance_str == "error" || balance_str == "not deployed" {
         return balance_str.to_string();
     }
 
     match balance_str.parse::<u128>() {
         Ok(balance) => {
-            let divisor = 10_u128.pow(decimals as u32);
+            let divisor = 10_u128.pow(decimals);
             let integer_part = balance / divisor;
             let fractional_part = balance % divisor;
